requirements.txt
alembic==1.14.1
aniso8601==10.0.0
attrs==25.1.0
click==8.1.8
Flask==2.2.5
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
cryptography==36.0.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
flask_mysql_connector==1.0.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
mistune==3.1.1
mysql-connector-python==9.2.0
numpy==2.2.2
pandas==2.2.3
PyJWT==2.10.1
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
